+++ packages/dart/test/package_test.dart
@@ -7,6 +7,6 @@ void main() {
   });
 
   test('包版本与 pubspec 一致', () {
-    expect(version, '0.1.0-beta.4');
+    expect(version, '0.1.0-beta.5');
   });
 }